% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kombinops.r
\name{Ank}
\alias{Ank}
\title{Count \emph{k}-permutations of \emph{n} elements (without repetition)}
\usage{
Ank(n, k)
}
\arguments{
\item{n}{The number of elements in the initial set.}

\item{k}{The number of elements in every resulting arrangement.}
}
\value{
The number of \code{k}-permutations of \code{n}
    (without repetition).
}
\description{
Count possible arrangements of a fixed length \code{k} of elements
taken from a given set of size \code{n}.
In other words, \code{k}-permutations of \code{n} (without repetition of elements from the initial set -
any element of the initial set can occure in every arrangement one or zero times).
The result is equal to zero when \code{k}>\code{n} and \eqn{ \frac{n!}{(n-k)!} }{ n! / (n-k)! } otherwise.
}
\details{
For 2-permutations of 3 without repetition we have 6 variants. If,
for example, the initial set is (1,2,3), then permutations will be:
(1,2), (1,3), (2,1), (2,3), (3,1) and (3,2).
}
\examples{
\dontrun{
Ank(2,3)
Ank(3,2)
Ank(length(c("l","u","c","k")),3)
}

}
