% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kombinops.r
\name{Ank_}
\alias{Ank_}
\title{Count \emph{k}-permutations of \emph{n} elements (with repetition)}
\usage{
Ank_(n, k)
}
\arguments{
\item{n}{The number of elements in the initial set.}

\item{k}{The number of elements in every resulting arrangement.}
}
\value{
The number of \code{k}-permutations of \code{n} (with possible repetition of
    elements from the initial set any number of times)
}
\description{
Count possible arrangements of a fixed length \code{k} of elements
taken from a given set of size \code{n}. In other words, \code{k}-permutations of \code{n}
(with possible repetition of elements from the initial set in every arrangment any number of times).
The result is equal to \eqn{ n^k }{ n^k }.
}
\details{
For 2-permutations of 3 with repetition we have 9 variants. If,
for example, the initial set is (1,2,3), then permitations will be:
(1,2), (1,3), (2,1), (2,3), (3,1), (3,2), (1,1), (2,2) and (3,3).
}
\examples{
\dontrun{
Ank_(2,3)
Ank_(3,2)
Ank_(length(c("l","u","c","k")),3)
}

}
