% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kombinops.r
\name{Cnk}
\alias{Cnk}
\title{Count \emph{k}-combinations of \emph{n} elements (without repetition)}
\usage{
Cnk(n, k)
}
\arguments{
\item{n}{The number of elements in the initial set.}

\item{k}{The number of elements in every resulting arrangement.}
}
\value{
The number of \code{k}-combinations of \code{n} (without repetition of elements
    from the initial set - any element of the initial set can occure in any
    combination one or zero times).
}
\description{
Count possible combinations of a fixed length \code{k} of elements
taken from a given set of size \code{n}. A combination is a selection of items
from a collection, such that (unlike permutations) the order of selection does not matter.
The result is equal to zero if \code{k}>\code{n} and \eqn{ \frac{n!}{k!*(n-k)!} }{ n! / k!(n-k)! }
otherwise.
}
\details{
For 2-combinations of 3 without repetition we have 3 variants. If,
for example, the initial set is (1,2,3), then combinations will be:
(1,2), (1,3) and (2,3).
}
\examples{
\dontrun{
Cnk(2,3)
Cnk(3,2)
Cnk(length(c("1","2","3","4")),3)
}

}
