% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kombinops.r
\name{Cnk_}
\alias{Cnk_}
\title{Count \emph{k}-combinations of \emph{n} elements (with repetition)}
\usage{
Cnk_(n, k)
}
\arguments{
\item{n}{The number of elements in the initial set.}

\item{k}{The number of elements in every resulting arrangement.}
}
\value{
The number of \code{k}-combinations of \code{n} (with repetition of elements
    from the initial set - any element of the initial set can occure in any
    combination any number of times).
}
\description{
Count possible combinations of a fixed length \code{k} of elements
taken from a given set of size \code{n}. A combination is a selection of items
from a collection, such that (unlike permutations) the order of selection does not matter.
The result is equal to \eqn{ \frac{(n+k-1)!}{k!(n-1)!} }{ (n+k-1)!/(k!*(n-1)!) }.
}
\details{
For 2-combinations of 3 with repetition we have 6 variants. If,
for example, the initial set is (1,2,3), then combinations will be:
(1,2), (1,3), (2,3), (1,1), (2,2) and (3,3).
}
\examples{
\dontrun{
Cnk_(2,3)
Cnk_(3,2)
Cnk_(length(c("1","2","3","4")),3)
}

}
