% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kombinops.r
\name{P}
\alias{P}
\title{Count permutations of multiset}
\usage{
P(nel)
}
\arguments{
\item{nel}{An integer vector of counts of each element in multiset.}
}
\value{
The number of permutations of multiset.
}
\description{
If \emph{M} is a finite multiset, then a multiset permutation is an ordered
arrangement of elements of \emph{M} in which each element appears exactly as
often as is its multiplicity in \emph{M}. An anagram of a word having some repeated
letters is an example of a multiset permutation. If the multiplicities of the elements of \emph{M}
(taken in some order) are \eqn{ m_1, m_2, ..., m_l }{ m1, m2, ..., ml } and
their sum (i.e., the size of \emph{M}) is \emph{n}, then the number of multiset permutations
of \emph{M} is given by the multinomial coefficient \eqn{ \frac{n!}{m_1!m_2!...m_l!} }{ n!/(m1!m2!...ml!) }.
}
\details{
For example, the number of distinct anagrams of the word MISSISSIPPI is
\eqn{ \frac{11!}{1!4!4!2!}=34650 }{ 11!/(1!4!4!2!)=34650 }.
}
\examples{
\dontrun{
# Permutations of the word "FERRARY"
letters<-strsplit("FERRARY", NULL)[[1]]
tw<-table(letters)
P(tw)
}

}
